Maybe fix weird Vim issue on some systems
Some select Linux remote systems act strangely in response to OS X
delete (backspace) and arrow keys. This is a shot in the dark, but
setting `term=cons25` appears to fix that with little or no side
effects.

I'm going to distribute this out to systems I touch and see if
keys work as expected.